rocjitsu: Simulation infra — dispatch, LDS, memory pipeline, FP16/packed ISA
- Enable SDMA queue processing and fix 2D/3D workgroup dispatch
- Fix LDS allocation collision and L2 write-through behavior
- Fix FP16 ISA correctness (v_cvt, packed ops) for vLLM kernel support
- Fix dot product, packed F32, and SDWA instruction implementations
- Implement v_pack_b32_f16 and fix cross-ISA sharing analysis
- Synchronous memory pipeline, ABI fixes, and allocation zeroing
- PackedTID workitem ID encoding for CDNA3/4
- CU count fix and code object metadata extraction
- SRAM ECC D16 zeroing behavior
- Add tmp/ to gitignore

+++ experimental/rocjitsu/lib/rocjitsu/src/rocjitsu/isa/arch/amdgpu/shared/addr_calc_flat.h
@@ -77,7 +77,8 @@ void flat_calculate_addresses(const FlatInst &inst, amdgpu::Wavefront &wf, Vecto
       uint32_t vbase = wf.vgpr_alloc().base + inst.addr;
       uint64_t vaddr;
       if (inst.saddr != 0x7F) {
-        vaddr = cu.read_vgpr(vbase, lane); // 32-bit VGPR offset
+        vaddr = static_cast<uint64_t>(static_cast<int64_t>(
+            static_cast<int32_t>(cu.read_vgpr(vbase, lane)))); // sign-extended 32-bit offset
       } else {
         vaddr = (static_cast<uint64_t>(cu.read_vgpr(vbase + 1, lane)) << 32) |
                 cu.read_vgpr(vbase, lane); // 64-bit VGPR pair
